Crate poseidon377

An instantiation of Poseidon for the BLS12-377 scalar field.

Modules

r1cs

Structs

Instance

Represents a generic instance of Poseidon.

PoseidonParameters

A set of Poseidon1 parameters for a given set of input parameters over decaf377::Fq.

Constants

RATE_1_PARAMS

Parameters for the rate-1 instance of Poseidon.

RATE_2_PARAMS

Parameters for the rate-2 instance of Poseidon.

RATE_3_PARAMS

Parameters for the rate-3 instance of Poseidon.

RATE_4_PARAMS

Parameters for the rate-4 instance of Poseidon.

RATE_5_PARAMS

Parameters for the rate-5 instance of Poseidon.

RATE_6_PARAMS

Parameters for the rate-6 instance of Poseidon.

RATE_7_PARAMS

Parameters for the rate-7 instance of Poseidon.

Functions

hash_1

Hash a single Fq element with the provided domain_separator.

hash_2

Hash two Fq elements with the provided domain_separator.

hash_3

Hash three Fq elements with the provided domain_separator.

hash_4

Hash four Fq elements with the provided domain_separator.

hash_5

Hash five Fq elements with the provided domain_separator.

hash_6

Hash six Fq elements with the provided domain_separator.

hash_7

Hash seven Fq elements with the provided domain_separator.

Type Aliases

Fq